+++ src/pages/playground/playground.js
@@ -1,208 +1,100 @@
-// components/Playground.tsx
 import React, { useState } from "react";
-import "./playground.css"
-import { RxCross1 } from "react-icons/rx";
-const Playground = () => {
-  const [collapsed, setSidebarCollapsed] = useState(false);
-  const navItems = [
-    {
-      label: "dashboard",
-      href: "/#!",
-      icon: <RxCross1 />,
-    },
-    {
-      label: "dashboard",
-      href: "/#!",
-      icon: <RxCross1 />,
-    },
-    {
-      label: "dashboard",
-      href: "/#!",
-      icon: <RxCross1 />,
-    },
-    {
-      label: "dashboard",
-      href: "/#!",
-      icon: <RxCross1 />,
-    },
-  ];
-  return (
-    <div
-      className={`grid min-h-screen ${
-        collapsed ? "grid-cols-sidebar-collapsed" : "grid-cols-sidebar"
-      } transition-[grid-template-columns] duration-300 ease-in-out`}
-    >
-      {/* sidebar */}
-      {/* https://reacthustle.com/blog/nextjs-react-responsive-collapsible-sidebar-tailwind#step-4-creating-the-sidebar-component */}
-      <div className="bg-green-700 text-white">
-        <button onClick={() => setSidebarCollapsed(!collapsed)}>close</button>
-        {!collapsed && <span className="whitespace-nowrap">My Logo</span>}
+import { toast } from "react-toastify";
 
-        {/* nav items part */}
-        <nav className="flex-grow">
-          <ul className={"my-2 flex flex-col gap-2 items-stretch"}>
-            {navItems.map((item, index) => {
-              return (
-                <li
-                  key={index}
-                  className={`text-green-100 hover:bg-green-900 flex transition-colors duration-300"
-                    ${
-                      collapsed
-                        ? "rounded-full p-2 mx-3 w-10 h-10"
-                        : "rounded-md p-2 mx-3 gap-4 "
-                    }`}
+const Playground = () => {
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+
+  const ProceedLogin = (e) => {
+    e.preventDefault();
+    if (validate()) {
+      console.log("go");
+    }
+  };
+
+  const validate = () => {
+    let result = true;
+    if (username === "" || username === null) {
+      result = false;
+      toast.error("Enter a username");
+    } else if (password === "" || password === null) {
+      result = false;
+      toast.error("Enter a password");
+    }
+    return result;
+  };
+
+  // https://www.youtube.com/watch?v=QapLfhqU8wY
+  return (
+    <section className="bg-[#efebff] dark:bg-gray-900">
+      <div className="flex flex-col items-center justify-center px-6 py-8 mx-auto md:h-screen lg:py-0">
+        <div className="w-full bg-white rounded-lg shadow dark:border md:mt-0 sm:max-w-md xl:p-0 dark:bg-gray-800 dark:border-gray-700">
+          <div className="p-6 space-y-4 md:space-y-6 sm:p-8">
+            <img
+              className="w-20 h-20 mx-auto object-contain"
+              src={require("../../assets/images/avatar.png")}
+              alt="logo"
+            />
+            <h1 className="text-xl text-center font-bold leading-tight tracking-tight text-gray-900 md:text-2xl dark:text-white">
+              Login
+            </h1>
+            <form
+              className="space-y-4 md:space-y-6"
+              action=""
+              onSubmit={ProceedLogin}
+            >
+              <div>
+                <label
+                  htmlFor="username"
+                  className="block mb-2 text-sm font-medium text-gray-900 dark:text-white"
                 >
-                  <a href={item.href} className="flex gap-2">
-                    {item.icon} <span>{!collapsed && item.label}</span>
-                  </a>
-                </li>
-              );
-            })}
-          </ul>
-        </nav>
+                  Username
+                </label>
+                <input
+                  type="text"
+                  name="username"
+                  id="username"
+                  className="bg-gray-50 border border-gray-300 text-gray-900 rounded-lg focus:ring-black focus:border-black block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                  placeholder="Enter Username"
+                  required=""
+                  value={username}
+                  onChange={(e) => {
+                    setUsername(e.target.value);
+                  }}
+                />
+              </div>
+              <div>
+                <label
+                  htmlFor="password"
+                  className="block mb-2 text-sm font-medium text-gray-900 dark:text-white"
+                >
+                  Password
+                </label>
+                <input
+                  type="password"
+                  name="password"
+                  id="password"
+                  placeholder="••••••••"
+                  className="bg-gray-50 border border-gray-300 text-gray-900 rounded-lg focus:ring-black focus:border-black block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                  required=""
+                  value={password}
+                  onChange={(e) => {
+                    setPassword(e.target.value);
+                  }}
+                />
+              </div>
+              <button
+                // type="submit"
+                className="w-full text-white bg-black hover:bg-primary-700 focus:ring-4 focus:outline-none focus:ring-primary-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-black dark:hover:bg-primary-700 dark:focus:ring-primary-800"
+              >
+                Sign in
+              </button>
+            </form>
+          </div>
+        </div>
       </div>
-      {/* content */}
-      <div className="">
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-        <p>
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Autem
-          possimus sequi, maiores delectus facilis saepe natus iure, officia
-          deserunt nostrum laudantium illum nihil quidem voluptatum ipsa eos
-          debitis ratione. Incidunt!
-        </p>
-      </div>
-    </div>
+    </section>
   );
 };
+
 export default Playground;